split melody into compiler and cli crates
