+++ crates/melody/cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "melody"
+version = "0.1.0"
+edition = "2021"
+rust-version = "1.58.0"
+
+[dependencies]
+logos = "0.12.0"
